requirements.txt
Cython==3.0.11
gitdb==4.0.11
GitPython==3.1.43
smmap==5.0.1
